+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 * Type of /TransformParams field in signature reference dictionary
+* [HexaPDF::Type::Page#box] to intersect the requested box with the media box
 
 
 ## 0.37.1 - 2024-02-05